Add files via upload
Introduction
The Consumer Complaint Database is a collection of complaints on a range of consumer financial products and services, sent to companies for response. It was designed to connect consumers with financial companies to understand issues with their mortgages, fix errors on their credit reports, stop harassment from debt collectors, and get direct responses about problems with their credit cards, checking and savings accounts, student loans, and more.
The objective of this database is, by analyzing the data, to identify trends and problems in the marketplace and to help to do a better job in supervising companies, enforcing federal consumer financial laws and writing rules and regulations.
•	Each complaint is forwarded to the appropriate company for a response.
•	The complaint data is shared with state and federal agencies. 
•	The complaint data is then analyzed to help with our work to supervise companies, enforce federal consumer financial laws, and write better rules and regulations.
Financial products and services sector have significant impact on our lives as they are directly related to our finances. It becomes need of hour to have proper understanding about how we can make correct decisions to buy a product or a service. Here in this project we analyze Customer complaints about various financial products to answer these questions.
Project Description
The objective of this project Complaint analysis is to track, categorize and handle customer complaints. When a customer registers a complaint, he or she is voicing a concern in relation to a product or service. However, not all complaints are to be treated equally and there are several questions to ask before taking action such as
•	Has this happened before?
•	Have the complaints been recorded?
•	How often does the same compliant arise?
•	Is there a pattern to this complaint in how it was received?
•	Has the same customer reported this previously?
1.	Analyzing the major customer concerns with banks in each state. We aim to gain an insight into the problems coming from retail banking in each state in the US.  This can be derived from the available dataset and the goal is to use classifying the customer concerns based on the state, will give the major concerns in each state and we aim to find the top 3 concerns of each customer.
2.	Derive business impact of customer concerns to banking institutions Classifying the available data based on whether customers' concerns were addressed or not will provide valuable insight into the effect of them on institutions' business. For example, a large number of unresolved complaints can be taken to mean that the specific customers have taken their business elsewhere.
3.	Develop a performance metric based on time taken to resolve concern Depending on the type of concern and the time taken to resolve it, we can come up with a standardized performance metric, taking into account the various factors of the case and apply the metric across all future solutions. This will help pinpoint. those problems that take longer to resolve and therefore can be given a higher priority.
4.	Perform the analysis to study the relationship of income with number of complaints.
5.	Study the impact of age, gender or household on complaints datasets in detail.
Techniques used
•	Hive aggregation queries 
•	HDFS
•	Pyspark
•	Graphs using aggregated data(optional)
•	Segmentation analysis using ML(optional)